SoundDriver

+++ RowDaBoat-x64barebones-d4e1c147f975/Kernel/interrupts/include/timerTick.h
@@ -1,9 +1,11 @@
 #ifndef __TIMER_TICK_H__
 #define __TIMER_TICK_H__
 
+#include <stdint.h>
+
 void timer_handler();
 int ticks_elapsed();
 int seconds_elapsed();
-void timer_wait(int time);
+void timer_wait(uint16_t time);
 
 #endif
